Cleanup, simplifying, and patching issues with the code
