+++ synapseclient/core/models/acl.py
@@ -32,46 +32,11 @@ class AclEntry:
     permissions: List[str] = field(default_factory=list)
     """List of permission strings granted to this principal."""
 
-    def has_permission(self, permission: str) -> bool:
-        """
-        Check if this ACL entry includes a specific permission.
-
-        Arguments:
-            permission: The permission string to check for (e.g., 'READ', 'DOWNLOAD').
-
-        Returns:
-            True if the permission is granted, False otherwise.
-        """
-        return permission in self.permissions
-
-    def add_permission(self, permission: str) -> None:
-        """
-        Add a permission to this ACL entry if not already present.
-
-        Arguments:
-            permission: The permission string to add.
-        """
-        if permission not in self.permissions:
-            self.permissions.append(permission)
-
-    def remove_permission(self, permission: str) -> None:
-        """
-        Remove a permission from this ACL entry if present.
-
-        Arguments:
-            permission: The permission string to remove.
-        """
-        if permission in self.permissions:
-            self.permissions.remove(permission)
-
 
 @dataclass
 class EntityAcl:
     """
     Represents the complete Access Control List for a single Synapse entity.
-
-    This dataclass contains all ACL entries for an entity, organizing them
-    by principal ID and providing convenient methods for ACL management.
 
     Attributes:
         entity_id: The Synapse ID of the entity (e.g., "syn123").
@@ -83,76 +48,6 @@ class EntityAcl:
 
     acl_entries: List[AclEntry] = field(default_factory=list)
     """List of ACL entries, each representing permissions for a principal."""
-
-    def get_acl_entry(self, principal_id: str) -> Optional[AclEntry]:
-        """
-        Get the ACL entry for a specific principal.
-
-        Arguments:
-            principal_id: The ID of the principal to look up.
-
-        Returns:
-            The AclEntry for the principal, or None if not found.
-        """
-        for entry in self.acl_entries:
-            if entry.principal_id == principal_id:
-                return entry
-        return None
-
-    def get_principals(self) -> List[str]:
-        """
-        Get a list of all principal IDs that have ACL entries for this entity.
-
-        Returns:
-            List of principal ID strings.
-        """
-        return [entry.principal_id for entry in self.acl_entries]
-
-    def get_permissions_for_principal(self, principal_id: str) -> List[str]:
-        """
-        Get the permissions list for a specific principal.
-
-        Arguments:
-            principal_id: The ID of the principal to look up.
-
-        Returns:
-            List of permission strings, or empty list if principal not found.
-        """
-        entry = self.get_acl_entry(principal_id)
-        return entry.permissions if entry else []
-
-    def add_acl_entry(self, acl_entry: AclEntry) -> None:
-        """
-        Add an ACL entry to this entity ACL.
-
-        If an entry for the same principal already exists, it will be replaced.
-
-        Arguments:
-            acl_entry: The ACL entry to add.
-        """
-        # Remove existing entry for the same principal if it exists
-        self.acl_entries = [
-            entry
-            for entry in self.acl_entries
-            if entry.principal_id != acl_entry.principal_id
-        ]
-        self.acl_entries.append(acl_entry)
-
-    def remove_acl_entry(self, principal_id: str) -> bool:
-        """
-        Remove the ACL entry for a specific principal.
-
-        Arguments:
-            principal_id: The ID of the principal whose ACL entry should be removed.
-
-        Returns:
-            True if an entry was removed, False if the principal was not found.
-        """
-        original_length = len(self.acl_entries)
-        self.acl_entries = [
-            entry for entry in self.acl_entries if entry.principal_id != principal_id
-        ]
-        return len(self.acl_entries) < original_length
 
     @classmethod
     def from_dict(cls, entity_id: str, acl_dict: Dict[str, List[str]]) -> "EntityAcl":
@@ -211,78 +106,9 @@ class AclListResult:
     entity_acls: List[EntityAcl] = field(default_factory=list)
     """List of EntityAcl objects, each representing the ACL for one entity."""
 
-    def get_entity_acl(self, entity_id: str) -> Optional[EntityAcl]:
-        """
-        Get the ACL for a specific entity.
-
-        Arguments:
-            entity_id: The Synapse ID of the entity to look up.
-
-        Returns:
-            The EntityAcl for the entity, or None if not found.
-        """
-        for entity_acl in self.entity_acls:
-            if entity_acl.entity_id == entity_id:
-                return entity_acl
-        return None
-
-    def get_entity_ids(self) -> List[str]:
-        """
-        Get a list of all entity IDs included in this ACL result.
-
-        Returns:
-            List of entity ID strings.
-        """
-        return [entity_acl.entity_id for entity_acl in self.entity_acls]
-
-    def get_permissions_for_entity_and_principal(
-        self, entity_id: str, principal_id: str
-    ) -> List[str]:
-        """
-        Get the permissions for a specific principal on a specific entity.
-
-        Arguments:
-            entity_id: The Synapse ID of the entity.
-            principal_id: The ID of the principal.
-
-        Returns:
-            List of permission strings, or empty list if entity or principal not found.
-        """
-        entity_acl = self.get_entity_acl(entity_id)
-        if entity_acl:
-            return entity_acl.get_permissions_for_principal(principal_id)
-        return []
-
-    def add_entity_acl(self, entity_acl: EntityAcl) -> None:
-        """
-        Add an EntityAcl to this result.
-
-        If an ACL for the same entity already exists, it will be replaced.
-
-        Arguments:
-            entity_acl: The EntityAcl to add.
-        """
-        # Remove existing ACL for the same entity if it exists
-        self.entity_acls = [
-            acl for acl in self.entity_acls if acl.entity_id != entity_acl.entity_id
-        ]
-        self.entity_acls.append(entity_acl)
-
-    def remove_entity_acl(self, entity_id: str) -> bool:
-        """
-        Remove the ACL for a specific entity.
-
-        Arguments:
-            entity_id: The Synapse ID of the entity whose ACL should be removed.
-
-        Returns:
-            True if an ACL was removed, False if the entity was not found.
-        """
-        original_length = len(self.entity_acls)
-        self.entity_acls = [
-            acl for acl in self.entity_acls if acl.entity_id != entity_id
-        ]
-        return len(self.entity_acls) < original_length
+    ascii_tree: Optional[str] = None
+    """Optional ASCII tree representation of the ACLs. This is only populated when
+    `log_tree` is set to True when calling `list_acl_async`."""
 
     @classmethod
     def from_dict(cls, acl_dict: Dict[str, Dict[str, List[str]]]) -> "AclListResult":
@@ -332,49 +158,3 @@ class AclListResult:
             entity_acl.entity_id: entity_acl.to_dict()
             for entity_acl in self.entity_acls
         }
-
-    def get_all_principals(self) -> List[str]:
-        """
-        Get a list of all unique principal IDs across all entities in this result.
-
-        Returns:
-            List of unique principal ID strings.
-        """
-        all_principals = set()
-        for entity_acl in self.entity_acls:
-            all_principals.update(entity_acl.get_principals())
-        return list(all_principals)
-
-    def get_entities_for_principal(self, principal_id: str) -> List[str]:
-        """
-        Get a list of entity IDs where the specified principal has any permissions.
-
-        Arguments:
-            principal_id: The ID of the principal to look up.
-
-        Returns:
-            List of entity ID strings where the principal has permissions.
-        """
-        entity_ids = []
-        for entity_acl in self.entity_acls:
-            if entity_acl.get_acl_entry(principal_id):
-                entity_ids.append(entity_acl.entity_id)
-        return entity_ids
-
-    def is_empty(self) -> bool:
-        """
-        Check if this result contains any ACL data.
-
-        Returns:
-            True if there are no entity ACLs, False otherwise.
-        """
-        return len(self.entity_acls) == 0
-
-    def __len__(self) -> int:
-        """
-        Get the number of entities in this ACL result.
-
-        Returns:
-            The number of entities with ACL data.
-        """
-        return len(self.entity_acls)
