+++ tests/integration/synapseclient/models/async/test_permissions_async.py
@@ -1,5 +1,6 @@
 """Integration tests for ACL on several models."""
 
+import asyncio
 import logging
 import uuid
 from typing import Callable, Dict, List, Optional, Type, Union
@@ -7,7 +8,9 @@ from typing import Callable, Dict, List, Optional, Type, Union
 import pytest
 
 from synapseclient import Synapse
+from synapseclient.client import AUTHENTICATED_USERS
 from synapseclient.core import utils
+from synapseclient.core.models.acl import AclListResult
 from synapseclient.models import (
     Column,
     ColumnType,
@@ -523,9 +526,15 @@ class TestDeletePermissions:
     """Test delete_permissions_async functionality across entities."""
 
     @pytest.fixture(autouse=True, scope="function")
-    def init(self, syn: Synapse, schedule_for_cleanup: Callable[..., None]) -> None:
+    def init(
+        self,
+        syn: Synapse,
+        syn_with_logger: Synapse,
+        schedule_for_cleanup: Callable[..., None],
+    ) -> None:
         self.syn = syn
         self.schedule_for_cleanup = schedule_for_cleanup
+        self.syn_with_logger = syn_with_logger
 
     @pytest.fixture(scope="function")
     def file(self, schedule_for_cleanup: Callable[..., None]) -> File:
@@ -638,213 +647,345 @@ class TestDeletePermissions:
         assert "READ" in acl
         return True
 
-    async def test_delete_permissions_single_file(
-        self, project_model: Project, file: File
-    ) -> None:
-        """Test deleting permissions on a single file."""
-        # GIVEN a file with custom permissions
-        file.name = f"test_file_{uuid.uuid4()}"
-        stored_file = await file.store_async(parent=project_model)
-        self.schedule_for_cleanup(stored_file.id)
-
-        await self._set_custom_permissions(stored_file)
-
-        # WHEN I delete permissions on the file
-        await stored_file.delete_permissions_async()
-
-        # THEN the permissions should be deleted
-        await self._verify_permissions_deleted(stored_file)
-
-    async def test_delete_permissions_single_folder(
-        self, project_model: Project
-    ) -> None:
-        """Test deleting permissions on a single folder."""
-        # GIVEN folder structure with permissions on top level folder
-        folder_structure = await self.create_folder_structure(project_model)
-        top_level_folder = folder_structure["top_level_folder"]
-
-        await self._set_custom_permissions(top_level_folder)
-
-        # WHEN I delete permissions on the folder
-        await top_level_folder.delete_permissions_async()
-
-        # THEN the permissions should be deleted
-        await self._verify_permissions_deleted(top_level_folder)
-
-    async def test_delete_permissions_skip_self(self, project_model: Project) -> None:
-        """Test deleting permissions with include_self=False."""
-        # GIVEN a folder structure with permissions set on the top level folder
-        folder_structure = await self.create_folder_structure(project_model)
-        top_level_folder = folder_structure["top_level_folder"]
-        files = folder_structure["files"]
-
-        # AND custom permissions are set on top level folder and a file
-        await self._set_custom_permissions(top_level_folder)
-        await self._set_custom_permissions(files[0])
-
-        # WHEN I delete permissions with include_self=False and include_container_content=True
-        await top_level_folder.delete_permissions_async(
-            include_self=False, include_container_content=True
+    async def _verify_list_acl_functionality(
+        self,
+        entity: Union[File, Folder, Project],
+        expected_entity_count: int,
+        recursive: bool = True,
+        include_container_content: bool = True,
+        target_entity_types: Optional[List[str]] = None,
+        log_tree: bool = True,
+    ) -> AclListResult:
+        """Helper to verify list_acl_async functionality and return results."""
+        await asyncio.sleep(2)
+        acl_result = await entity.list_acl_async(
+            recursive=recursive,
+            include_container_content=include_container_content,
+            target_entity_types=target_entity_types,
+            log_tree=log_tree,
+            synapse_client=self.syn_with_logger,
         )
 
-        # THEN the top level folder permissions should remain
-        assert await self._verify_permissions_not_deleted(top_level_folder)
+        assert isinstance(acl_result, AclListResult)
+        assert len(acl_result.entity_acls) >= expected_entity_count
 
-        # AND the file permissions should be deleted
-        await self._verify_permissions_deleted(files[0])
+        return acl_result
 
-    async def test_delete_permissions_include_container_content(
-        self, project_model: Project
+    def _verify_log_messages(
+        self,
+        caplog: pytest.LogCaptureFixture,
+        list_acl_called: bool = True,
+        delete_permissions_called: bool = True,
+        dry_run: bool = False,
+        tree_logging: bool = True,
     ) -> None:
-        """Test deleting permissions with include_container_content=True."""
-        # GIVEN a folder structure with permissions set on top level folder and files
-        folder_structure = await self.create_folder_structure(project_model)
-        top_level_folder = folder_structure["top_level_folder"]
-        files = folder_structure["files"]
-        folder_1 = folder_structure["folder_1"]
+        """Helper to verify expected log messages from both methods."""
+        log_text = caplog.text
 
-        # AND permissions are set on top level folder, files, and folder_1
-        await self._set_custom_permissions(top_level_folder)
-        await self._set_custom_permissions(files[0])
-        await self._set_custom_permissions(folder_1)
+        if list_acl_called and tree_logging:
+            assert "ACL Tree Structure:" in log_text
 
-        # WHEN I delete permissions with include_container_content=True but recursive=False
-        await top_level_folder.delete_permissions_async(
-            include_self=True, include_container_content=True, recursive=False
+        if delete_permissions_called and dry_run:
+            assert "DRY RUN" in log_text
+            assert "Permission Deletion Impact Analysis" in log_text
+            assert "End of Dry Run Analysis" in log_text
+
+    async def create_simple_tree_structure(
+        self, project_model: Project
+    ) -> Dict[str, Union[Folder, File]]:
+        """
+        Create a simple 2-level tree structure.
+
+        Structure:
+        ```
+        Project
+        └── folder_a
+            └── file_1
+        ```
+        """
+        folder_a = await Folder(name=f"folder_a_{uuid.uuid4()}").store_async(
+            parent=project_model, synapse_client=self.syn
+        )
+        self.schedule_for_cleanup(folder_a.id)
+
+        file_1 = await File(
+            path=utils.make_bogus_uuid_file(), name=f"file_1_{uuid.uuid4()}"
+        ).store_async(parent=folder_a, synapse_client=self.syn)
+        self.schedule_for_cleanup(file_1.id)
+
+        return {
+            "folder_a": folder_a,
+            "file_1": file_1,
+        }
+
+    async def create_deep_nested_structure(
+        self, project_model: Project
+    ) -> Dict[str, Union[Folder, File]]:
+        """
+        Create a deeply nested folder structure with files at various levels.
+
+        Structure:
+        ```
+        Project
+        └── level_1
+            ├── file_at_1
+            └── level_2
+                ├── file_at_2
+                └── level_3
+                    ├── file_at_3
+                    └── level_4
+                        └── file_at_4
+        ```
+        """
+        level_1 = await Folder(name=f"level_1_{uuid.uuid4()}").store_async(
+            parent=project_model, synapse_client=self.syn
+        )
+        self.schedule_for_cleanup(level_1.id)
+
+        # Create file_at_1 and level_2 in parallel since they don't depend on each other
+        file_at_1_task = File(
+            path=utils.make_bogus_uuid_file(), name=f"file_at_1_{uuid.uuid4()}"
+        ).store_async(parent=level_1, synapse_client=self.syn)
+        level_2_task = Folder(name=f"level_2_{uuid.uuid4()}").store_async(
+            parent=level_1, synapse_client=self.syn
         )
 
-        # THEN the top level folder permissions should be deleted
-        await self._verify_permissions_deleted(top_level_folder)
+        file_at_1, level_2 = await asyncio.gather(file_at_1_task, level_2_task)
+        self.schedule_for_cleanup(file_at_1.id)
+        self.schedule_for_cleanup(level_2.id)
 
-        # AND the files permissions should be deleted
-        await self._verify_permissions_deleted(files[0])
+        # Create file_at_2 and level_3 in parallel since they don't depend on each other
+        file_at_2_task = File(
+            path=utils.make_bogus_uuid_file(), name=f"file_at_2_{uuid.uuid4()}"
+        ).store_async(parent=level_2, synapse_client=self.syn)
+        level_3_task = Folder(name=f"level_3_{uuid.uuid4()}").store_async(
+            parent=level_2, synapse_client=self.syn
+        )
 
-        # AND the folder_1 permissions should be deleted
-        await self._verify_permissions_deleted(folder_1)
+        file_at_2, level_3 = await asyncio.gather(file_at_2_task, level_3_task)
+        self.schedule_for_cleanup(file_at_2.id)
+        self.schedule_for_cleanup(level_3.id)
 
-    async def test_delete_permissions_recursive(self, project_model: Project) -> None:
-        """Test deleting permissions recursively."""
-        # GIVEN a folder structure with permissions set throughout
-        folder_structure = await self.create_folder_structure(project_model)
-        top_level_folder = folder_structure["top_level_folder"]
-        folder_1 = folder_structure["folder_1"]
-        file_3 = folder_structure["file_3"]
+        # Create file_at_3 and level_4 in parallel since they don't depend on each other
+        file_at_3_task = File(
+            path=utils.make_bogus_uuid_file(), name=f"file_at_3_{uuid.uuid4()}"
+        ).store_async(parent=level_3, synapse_client=self.syn)
+        level_4_task = Folder(name=f"level_4_{uuid.uuid4()}").store_async(
+            parent=level_3, synapse_client=self.syn
+        )
 
-        # AND permissions are set on top_level_folder, folder_1, and file_3
-        await self._set_custom_permissions(top_level_folder)
-        await self._set_custom_permissions(folder_1)
-        await self._set_custom_permissions(file_3)
+        file_at_3, level_4 = await asyncio.gather(file_at_3_task, level_4_task)
+        self.schedule_for_cleanup(file_at_3.id)
+        self.schedule_for_cleanup(level_4.id)
 
-        # WHEN I delete permissions recursively but without container content
-        with pytest.raises(
-            ValueError,
-            match="When recursive=True, include_container_content must also be True",
-        ) as exc_info:
-            await top_level_folder.delete_permissions_async(
-                recursive=True, include_container_content=False
+        file_at_4 = await File(
+            path=utils.make_bogus_uuid_file(), name=f"file_at_4_{uuid.uuid4()}"
+        ).store_async(parent=level_4, synapse_client=self.syn)
+        self.schedule_for_cleanup(file_at_4.id)
+
+        return {
+            "level_1": level_1,
+            "level_2": level_2,
+            "level_3": level_3,
+            "level_4": level_4,
+            "file_at_1": file_at_1,
+            "file_at_2": file_at_2,
+            "file_at_3": file_at_3,
+            "file_at_4": file_at_4,
+        }
+
+    async def create_wide_tree_structure(
+        self, project_model: Project
+    ) -> Dict[str, Union[Folder, List[Union[Folder, File]]]]:
+        """
+        Create a wide tree structure with multiple siblings.
+
+        Structure:
+        ```
+        Project
+        ├── folder_a
+        │   └── file_a
+        ├── folder_b
+        │   └── file_b
+        ├── folder_c
+        │   └── file_c
+        └── root_file
+        ```
+        """
+        # Create folders in parallel
+        folder_tasks = [
+            Folder(name=f"folder_{folder_letter}_{uuid.uuid4()}").store_async(
+                parent=project_model, synapse_client=self.syn
             )
+            for folder_letter in ["a", "b", "c"]
+        ]
+        folders = await asyncio.gather(*folder_tasks)
 
-        # THEN the top_level_folder permissions should be deleted
-        await self._verify_permissions_deleted(top_level_folder)
+        # Schedule cleanup for folders
+        for folder in folders:
+            self.schedule_for_cleanup(folder.id)
 
-        # AND the folder_1 permissions should remain (because include_container_content=False)
-        await self._verify_permissions_not_deleted(folder_1)
+        # Create files in parallel
+        file_tasks = [
+            File(
+                path=utils.make_bogus_uuid_file(),
+                name=f"file_{folder_letter}_{uuid.uuid4()}",
+            ).store_async(parent=folder, synapse_client=self.syn)
+            for folder_letter, folder in zip(["a", "b", "c"], folders)
+        ]
 
-        # AND the file_3 permissions should remain (because include_container_content=False)
-        assert await self._verify_permissions_not_deleted(file_3)
+        # Create root file task
+        root_file_task = File(
+            path=utils.make_bogus_uuid_file(), name=f"root_file_{uuid.uuid4()}"
+        ).store_async(parent=project_model, synapse_client=self.syn)
 
-    async def test_delete_permissions_recursive_with_container_content(
+        # Execute file creation tasks in parallel
+        file_results = await asyncio.gather(*file_tasks, root_file_task)
+        all_files = file_results[:-1]  # All but the last (root file)
+        root_file = file_results[-1]  # The last one (root file)
+
+        # Schedule cleanup for files
+        for file in all_files:
+            self.schedule_for_cleanup(file.id)
+        self.schedule_for_cleanup(root_file.id)
+
+        return {
+            "folders": folders,
+            "all_files": all_files,
+            "root_file": root_file,
+        }
+
+    async def create_complex_mixed_structure(
         self, project_model: Project
-    ) -> None:
-        """Test deleting permissions recursively with container content."""
-        # GIVEN a folder structure with permissions set throughout
-        folder_structure = await self.create_folder_structure(project_model)
-        top_level_folder = folder_structure["top_level_folder"]
-        files = folder_structure["files"]
-        folder_1 = folder_structure["folder_1"]
-        file_3 = folder_structure["file_3"]
+    ) -> Dict[str, Union[Folder, File, List]]:
+        """
+        Create a complex mixed structure combining depth and width.
 
-        # AND permissions are set on top_level_folder, files, folder_1, and file_3
-        await self._set_custom_permissions(top_level_folder)
-        await self._set_custom_permissions(files[0])
-        await self._set_custom_permissions(folder_1)
-        await self._set_custom_permissions(file_3)
-
-        # WHEN I delete permissions recursively with container content
-        await top_level_folder.delete_permissions_async(
-            recursive=True, include_container_content=True
+        Structure:
+        ```
+        Project
+        ├── shallow_folder
+        │   └── shallow_file
+        ├── deep_branch
+        │   ├── deep_file_1
+        │   └── sub_deep
+        │       ├── deep_file_2
+        │       └── sub_sub_deep
+        │           └── deep_file_3
+        └── mixed_folder
+            ├── mixed_file
+            ├── mixed_sub_a
+            │   └── mixed_file_a
+            └── mixed_sub_b
+                └── mixed_file_b
+        ```
+        """
+        # Create top-level folders in parallel
+        top_folder_tasks = [
+            Folder(name=f"shallow_folder_{uuid.uuid4()}").store_async(
+                parent=project_model, synapse_client=self.syn
+            ),
+            Folder(name=f"deep_branch_{uuid.uuid4()}").store_async(
+                parent=project_model, synapse_client=self.syn
+            ),
+            Folder(name=f"mixed_folder_{uuid.uuid4()}").store_async(
+                parent=project_model, synapse_client=self.syn
+            ),
+        ]
+        shallow_folder, deep_branch, mixed_folder = await asyncio.gather(
+            *top_folder_tasks
         )
 
-        # THEN all permissions should be deleted
-        await self._verify_permissions_deleted(top_level_folder)
-        await self._verify_permissions_deleted(files[0])
-        await self._verify_permissions_deleted(folder_1)
-        await self._verify_permissions_deleted(file_3)
+        # Schedule cleanup for top-level folders
+        for folder in [shallow_folder, deep_branch, mixed_folder]:
+            self.schedule_for_cleanup(folder.id)
 
-    async def test_delete_permissions_target_entity_types_folder_only(
-        self, project_model: Project
-    ) -> None:
-        """Test deleting permissions with target_entity_types=['folder']."""
-        # GIVEN a folder structure with permissions set throughout
-        folder_structure = await self.create_folder_structure(project_model)
-        top_level_folder = folder_structure["top_level_folder"]
-        files = folder_structure["files"]
-        folder_1 = folder_structure["folder_1"]
-        sub_folder_1 = folder_structure["sub_folder_1"]
+        # Create first level files and folders
+        shallow_file = await File(
+            path=utils.make_bogus_uuid_file(), name=f"shallow_file_{uuid.uuid4()}"
+        ).store_async(parent=shallow_folder, synapse_client=self.syn)
+        self.schedule_for_cleanup(shallow_file.id)
 
-        # AND permissions are set on top_level_folder, files, folder_1, and sub_folder_1
-        await self._set_custom_permissions(top_level_folder)
-        await self._set_custom_permissions(files[0])
-        await self._set_custom_permissions(folder_1)
-        await self._set_custom_permissions(sub_folder_1)
+        # Deep branch structure
+        deep_file_1_task = File(
+            path=utils.make_bogus_uuid_file(), name=f"deep_file_1_{uuid.uuid4()}"
+        ).store_async(parent=deep_branch, synapse_client=self.syn)
 
-        # WHEN I delete permissions recursively but only for folder entity types
-        await top_level_folder.delete_permissions_async(
-            recursive=True,
-            include_container_content=True,
-            target_entity_types=["folder"],
+        sub_deep_task = Folder(name=f"sub_deep_{uuid.uuid4()}").store_async(
+            parent=deep_branch, synapse_client=self.syn
         )
 
-        # THEN folder permissions should be deleted
-        await self._verify_permissions_deleted(top_level_folder)
-        await self._verify_permissions_deleted(folder_1)
-        await self._verify_permissions_deleted(sub_folder_1)
+        deep_file_1, sub_deep = await asyncio.gather(deep_file_1_task, sub_deep_task)
+        self.schedule_for_cleanup(deep_file_1.id)
+        self.schedule_for_cleanup(sub_deep.id)
 
-        # BUT file permissions should remain
-        assert await self._verify_permissions_not_deleted(files[0])
+        # Continue deep structure
+        deep_file_2_task = File(
+            path=utils.make_bogus_uuid_file(), name=f"deep_file_2_{uuid.uuid4()}"
+        ).store_async(parent=sub_deep, synapse_client=self.syn)
 
-    async def test_delete_permissions_target_entity_types_file_only(
-        self, project_model: Project
-    ) -> None:
-        """Test deleting permissions with target_entity_types=['file']."""
-        # GIVEN a folder structure with permissions set throughout
-        folder_structure = await self.create_folder_structure(project_model)
-        top_level_folder = folder_structure["top_level_folder"]
-        files = folder_structure["files"]
-        folder_1 = folder_structure["folder_1"]
-        file_3 = folder_structure["file_3"]
-        file_4 = folder_structure["file_4"]
-
-        # AND permissions are set on top_level_folder, files, folder_1, file_3, and file_4
-        await self._set_custom_permissions(top_level_folder)
-        await self._set_custom_permissions(files[0])
-        await self._set_custom_permissions(folder_1)
-        await self._set_custom_permissions(file_3)
-        await self._set_custom_permissions(file_4)
-
-        # WHEN I delete permissions recursively but only for file entity types
-        await top_level_folder.delete_permissions_async(
-            recursive=True, include_container_content=True, target_entity_types=["file"]
+        sub_sub_deep_task = Folder(name=f"sub_sub_deep_{uuid.uuid4()}").store_async(
+            parent=sub_deep, synapse_client=self.syn
         )
 
-        # THEN file permissions should be deleted
-        await self._verify_permissions_deleted(files[0])
-        await self._verify_permissions_deleted(file_3)
-        await self._verify_permissions_deleted(file_4)
+        deep_file_2, sub_sub_deep = await asyncio.gather(
+            deep_file_2_task, sub_sub_deep_task
+        )
+        self.schedule_for_cleanup(deep_file_2.id)
+        self.schedule_for_cleanup(sub_sub_deep.id)
 
-        # BUT folder permissions should remain
-        assert await self._verify_permissions_not_deleted(top_level_folder)
-        assert await self._verify_permissions_not_deleted(folder_1)
+        deep_file_3 = await File(
+            path=utils.make_bogus_uuid_file(), name=f"deep_file_3_{uuid.uuid4()}"
+        ).store_async(parent=sub_sub_deep, synapse_client=self.syn)
+        self.schedule_for_cleanup(deep_file_3.id)
+
+        # Mixed folder structure
+        mixed_file_task = File(
+            path=utils.make_bogus_uuid_file(), name=f"mixed_file_{uuid.uuid4()}"
+        ).store_async(parent=mixed_folder, synapse_client=self.syn)
+
+        mixed_sub_a_task = Folder(name=f"mixed_sub_a_{uuid.uuid4()}").store_async(
+            parent=mixed_folder, synapse_client=self.syn
+        )
+        mixed_sub_b_task = Folder(name=f"mixed_sub_b_{uuid.uuid4()}").store_async(
+            parent=mixed_folder, synapse_client=self.syn
+        )
+
+        mixed_file, mixed_sub_a, mixed_sub_b = await asyncio.gather(
+            mixed_file_task, mixed_sub_a_task, mixed_sub_b_task
+        )
+
+        # Schedule cleanup
+        self.schedule_for_cleanup(mixed_file.id)
+        self.schedule_for_cleanup(mixed_sub_a.id)
+        self.schedule_for_cleanup(mixed_sub_b.id)
+
+        # Create files in mixed sub-folders in parallel
+        mixed_file_a_task = File(
+            path=utils.make_bogus_uuid_file(), name=f"mixed_file_a_{uuid.uuid4()}"
+        ).store_async(parent=mixed_sub_a, synapse_client=self.syn)
+
+        mixed_file_b_task = File(
+            path=utils.make_bogus_uuid_file(), name=f"mixed_file_b_{uuid.uuid4()}"
+        ).store_async(parent=mixed_sub_b, synapse_client=self.syn)
+
+        mixed_file_a, mixed_file_b = await asyncio.gather(
+            mixed_file_a_task, mixed_file_b_task
+        )
+        self.schedule_for_cleanup(mixed_file_a.id)
+        self.schedule_for_cleanup(mixed_file_b.id)
+
+        return {
+            "shallow_folder": shallow_folder,
+            "shallow_file": shallow_file,
+            "deep_branch": deep_branch,
+            "sub_deep": sub_deep,
+            "sub_sub_deep": sub_sub_deep,
+            "deep_files": [deep_file_1, deep_file_2, deep_file_3],
+            "mixed_folder": mixed_folder,
+            "mixed_sub_folders": [mixed_sub_a, mixed_sub_b],
+            "mixed_files": [mixed_file, mixed_file_a, mixed_file_b],
+        }
 
     async def test_delete_permissions_invalid_entity_type(
         self, project_model: Project
@@ -878,6 +1019,7 @@ class TestDeletePermissions:
 
         # AND custom permissions are set for authenticated users
         await self._set_custom_permissions(project)
+        await asyncio.sleep(2)
 
         # WHEN I delete permissions on the project
         await project.delete_permissions_async()
@@ -894,3 +1036,983 @@ class TestDeletePermissions:
             # Alternatively, verify that the permissions weren't actually deleted
             # by checking if they still exist
             assert await self._verify_permissions_not_deleted(project)
+
+    async def test_delete_permissions_simple_tree_structure(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test deleting permissions on a simple tree structure."""
+        # GIVEN a simple tree structure with permissions
+        structure = await self.create_simple_tree_structure(project_model)
+        folder_a = structure["folder_a"]
+        file_1 = structure["file_1"]
+
+        # Set permissions on all entities
+        await asyncio.gather(
+            self._set_custom_permissions(folder_a),
+            self._set_custom_permissions(file_1),
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async before deletion
+        await self._verify_list_acl_functionality(
+            entity=folder_a,
+            expected_entity_count=2,  # folder_a and file_1
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()  # Clear logs for next verification
+
+        # WHEN I delete permissions recursively
+        await folder_a.delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN all permissions should be deleted
+        await asyncio.gather(
+            self._verify_permissions_deleted(folder_a),
+            self._verify_permissions_deleted(file_1),
+        )
+
+    async def test_delete_permissions_deep_nested_structure(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test deleting permissions on a deeply nested structure."""
+        # GIVEN a deeply nested structure with permissions
+        structure = await self.create_deep_nested_structure(project_model)
+
+        # Set permissions on all entities
+        await asyncio.gather(
+            *[self._set_custom_permissions(entity) for entity in structure.values()]
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async before deletion
+        await self._verify_list_acl_functionality(
+            entity=structure["level_1"],
+            expected_entity_count=8,  # all levels and files
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions recursively from the top level
+        await structure["level_1"].delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN all permissions should be deleted
+        await asyncio.gather(
+            *[self._verify_permissions_deleted(entity) for entity in structure.values()]
+        )
+
+    async def test_delete_permissions_wide_tree_structure(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test deleting permissions on a wide tree structure with multiple siblings."""
+        # GIVEN a wide tree structure with permissions
+        structure = await self.create_wide_tree_structure(project_model)
+        folders = structure["folders"]
+        all_files = structure["all_files"]
+        root_file = structure["root_file"]
+
+        # Set permissions on all entities
+        entities_to_set = folders + all_files + [root_file]
+        await asyncio.gather(
+            *[self._set_custom_permissions(entity) for entity in entities_to_set]
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async before deletion
+        await self._verify_list_acl_functionality(
+            entity=project_model,
+            expected_entity_count=7,  # 3 folders + 3 files + 1 root file
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions recursively from the project
+        await project_model.delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN all permissions should be deleted (except project which can't be deleted)
+        entities_to_verify = folders + all_files + [root_file]
+        await asyncio.gather(
+            *[self._verify_permissions_deleted(entity) for entity in entities_to_verify]
+        )
+
+    async def test_delete_permissions_complex_mixed_structure(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test deleting permissions on a complex mixed structure."""
+        # GIVEN a complex mixed structure with permissions
+        structure = await self.create_complex_mixed_structure(project_model)
+
+        # Set permissions on all entities
+        entities_to_set = (
+            [
+                structure["shallow_folder"],
+                structure["shallow_file"],
+                structure["deep_branch"],
+                structure["sub_deep"],
+                structure["sub_sub_deep"],
+                structure["mixed_folder"],
+            ]
+            + structure["deep_files"]
+            + structure["mixed_sub_folders"]
+            + structure["mixed_files"]
+        )
+
+        await asyncio.gather(
+            *[self._set_custom_permissions(entity) for entity in entities_to_set]
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async before deletion
+        await self._verify_list_acl_functionality(
+            entity=project_model,
+            expected_entity_count=12,  # complex structure with multiple entities
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions recursively from the project
+        await project_model.delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN all permissions should be deleted
+        await asyncio.gather(
+            *[self._verify_permissions_deleted(entity) for entity in entities_to_set]
+        )
+
+    # Edge case tests
+    async def test_delete_permissions_empty_folder(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test deleting permissions on an empty folder."""
+        # GIVEN an empty folder with custom permissions
+        empty_folder = await Folder(name=f"empty_folder_{uuid.uuid4()}").store_async(
+            parent=project_model, synapse_client=self.syn
+        )
+        self.schedule_for_cleanup(empty_folder.id)
+        await self._set_custom_permissions(empty_folder)
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async before deletion (empty folder)
+        await self._verify_list_acl_functionality(
+            entity=empty_folder,
+            expected_entity_count=1,  # just the empty folder itself
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred for empty structure
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions recursively
+        await empty_folder.delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN the folder permissions should be deleted
+        await self._verify_permissions_deleted(empty_folder)
+
+    async def test_delete_permissions_folder_with_only_files(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test deleting permissions on a folder that contains only files."""
+        # GIVEN a folder with only one file
+        folder = await Folder(name=f"files_only_folder_{uuid.uuid4()}").store_async(
+            parent=project_model, synapse_client=self.syn
+        )
+        self.schedule_for_cleanup(folder.id)
+
+        file = await File(
+            path=utils.make_bogus_uuid_file(), name=f"only_file_{uuid.uuid4()}"
+        ).store_async(parent=folder, synapse_client=self.syn)
+        self.schedule_for_cleanup(file.id)
+
+        # Set permissions on all entities
+        await asyncio.gather(
+            self._set_custom_permissions(folder),
+            self._set_custom_permissions(file),
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async before deletion
+        await self._verify_list_acl_functionality(
+            entity=folder,
+            expected_entity_count=2,  # folder and file
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions recursively
+        await folder.delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN all permissions should be deleted
+        await asyncio.gather(
+            self._verify_permissions_deleted(folder),
+            self._verify_permissions_deleted(file),
+        )
+
+    async def test_delete_permissions_folder_with_only_folders(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test deleting permissions on a folder that contains only sub-folders."""
+        # GIVEN a folder with only sub-folders
+        parent_folder = await Folder(
+            name=f"folders_only_parent_{uuid.uuid4()}"
+        ).store_async(parent=project_model, synapse_client=self.syn)
+        self.schedule_for_cleanup(parent_folder.id)
+
+        # Create sub-folders in parallel
+        sub_folder_tasks = [
+            Folder(name=f"only_subfolder_{i}_{uuid.uuid4()}").store_async(
+                parent=parent_folder, synapse_client=self.syn
+            )
+            for i in range(3)
+        ]
+        sub_folders = await asyncio.gather(*sub_folder_tasks)
+
+        # Schedule cleanup for sub-folders
+        for sub_folder in sub_folders:
+            self.schedule_for_cleanup(sub_folder.id)
+
+        # Set permissions on all entities
+        entities_to_set = [parent_folder] + sub_folders
+        await asyncio.gather(
+            *[self._set_custom_permissions(entity) for entity in entities_to_set]
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async before deletion
+        await self._verify_list_acl_functionality(
+            entity=parent_folder,
+            expected_entity_count=4,  # parent + 3 sub-folders
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions recursively
+        await parent_folder.delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN all permissions should be deleted
+        await asyncio.gather(
+            *[self._verify_permissions_deleted(entity) for entity in entities_to_set]
+        )
+
+    async def test_delete_permissions_target_files_only_complex(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test deleting permissions targeting only files in a complex structure."""
+        # GIVEN a complex structure with permissions
+        structure = await self.create_complex_mixed_structure(project_model)
+
+        # Set permissions on all entities
+        await asyncio.gather(
+            self._set_custom_permissions(structure["shallow_folder"]),
+            self._set_custom_permissions(structure["shallow_file"]),
+            self._set_custom_permissions(structure["deep_branch"]),
+            self._set_custom_permissions(structure["sub_deep"]),
+            *[self._set_custom_permissions(file) for file in structure["deep_files"]],
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async with target_entity_types for files only
+        await self._verify_list_acl_functionality(
+            entity=project_model,
+            expected_entity_count=4,  # shallow_file + 3 deep_files
+            recursive=True,
+            include_container_content=True,
+            target_entity_types=["file"],
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred with file filtering
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions targeting only files
+        await project_model.delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            target_entity_types=["file"],
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN only file permissions should be deleted
+        await asyncio.gather(
+            self._verify_permissions_deleted(structure["shallow_file"]),
+            *[
+                self._verify_permissions_deleted(file)
+                for file in structure["deep_files"]
+            ],
+        )
+
+        # BUT folder permissions should remain
+        await asyncio.gather(
+            self._verify_permissions_not_deleted(structure["shallow_folder"]),
+            self._verify_permissions_not_deleted(structure["deep_branch"]),
+            self._verify_permissions_not_deleted(structure["sub_deep"]),
+        )
+
+    # Include container content vs recursive tests
+    async def test_delete_permissions_include_container_only_deep_structure(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test include_container_content=True without recursive on deep structure."""
+        # GIVEN a deep nested structure with permissions
+        structure = await self.create_deep_nested_structure(project_model)
+
+        # Set permissions on all entities
+        await asyncio.gather(
+            *[self._set_custom_permissions(entity) for entity in structure.values()]
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async with include_container_content=True
+        await self._verify_list_acl_functionality(
+            entity=structure["level_1"],
+            expected_entity_count=3,  # level_1, file_at_1, level_2 (non-recursive)
+            recursive=False,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions with include_container_content=True but recursive=False
+        await structure["level_1"].delete_permissions_async(
+            include_self=True,
+            include_container_content=True,
+            recursive=False,
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN only level_1 and its direct children should have permissions deleted
+        await asyncio.gather(
+            self._verify_permissions_deleted(structure["level_1"]),
+            self._verify_permissions_deleted(structure["file_at_1"]),
+            self._verify_permissions_deleted(structure["level_2"]),
+        )
+
+        # BUT deeper nested entities should retain permissions
+        await asyncio.gather(
+            self._verify_permissions_not_deleted(structure["level_3"]),
+            self._verify_permissions_not_deleted(structure["level_4"]),
+            self._verify_permissions_not_deleted(structure["file_at_2"]),
+            self._verify_permissions_not_deleted(structure["file_at_3"]),
+            self._verify_permissions_not_deleted(structure["file_at_4"]),
+        )
+
+    async def test_delete_permissions_skip_self_complex_structure(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test include_self=False on a complex structure."""
+        # GIVEN a complex mixed structure with permissions
+        structure = await self.create_complex_mixed_structure(project_model)
+
+        # Set permissions on all entities
+        await asyncio.gather(
+            self._set_custom_permissions(structure["mixed_folder"]),
+            *[
+                self._set_custom_permissions(folder)
+                for folder in structure["mixed_sub_folders"]
+            ],
+            *[self._set_custom_permissions(file) for file in structure["mixed_files"]],
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async before deletion (should show all entities)
+        await self._verify_list_acl_functionality(
+            entity=structure["mixed_folder"],
+            expected_entity_count=4,  # mixed_folder + 2 sub_folders + 1 mixed_file
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions with include_self=False
+        await structure["mixed_folder"].delete_permissions_async(
+            include_self=False,
+            include_container_content=True,
+            recursive=True,
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN the mixed_folder permissions should remain
+        await self._verify_permissions_not_deleted(structure["mixed_folder"])
+
+        # BUT child permissions should be deleted
+        await asyncio.gather(
+            *[
+                self._verify_permissions_deleted(folder)
+                for folder in structure["mixed_sub_folders"]
+            ],
+            *[
+                self._verify_permissions_deleted(file)
+                for file in structure["mixed_files"]
+            ],
+        )
+
+    # Dry run functionality tests
+    async def test_delete_permissions_dry_run_no_changes(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test that dry_run=True makes no actual changes."""
+        # GIVEN a simple structure with permissions
+        structure = await self.create_simple_tree_structure(project_model)
+        folder_a = structure["folder_a"]
+        file_1 = structure["file_1"]
+
+        # Set permissions on all entities
+        await asyncio.gather(
+            self._set_custom_permissions(folder_a),
+            self._set_custom_permissions(file_1),
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async before dry run
+        initial_acl_result = await self._verify_list_acl_functionality(
+            entity=folder_a,
+            expected_entity_count=2,  # folder_a and file_1
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I run delete_permissions with dry_run=True
+        await folder_a.delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            dry_run=True,
+            synapse_client=self.syn_with_logger,
+        )
+
+        # THEN no permissions should be deleted
+        await asyncio.gather(
+            self._verify_permissions_not_deleted(folder_a),
+            self._verify_permissions_not_deleted(file_1),
+        )
+
+        # AND dry run messages should be logged
+        self._verify_log_messages(
+            caplog,
+            list_acl_called=False,
+            delete_permissions_called=True,
+            dry_run=True,
+            tree_logging=False,
+        )
+
+        # WHEN - Verify list_acl_async after dry run (should be identical)
+        caplog.clear()
+        final_acl_result = await self._verify_list_acl_functionality(
+            entity=folder_a,
+            expected_entity_count=2,  # should be same as before
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Results should be identical since dry run made no changes
+        assert len(initial_acl_result.entity_acls) == len(final_acl_result.entity_acls)
+
+    async def test_delete_permissions_dry_run_complex_logging(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test dry run logging for complex structures."""
+        # GIVEN a complex structure with permissions
+        structure = await self.create_complex_mixed_structure(project_model)
+
+        # Set permissions on a subset of entities
+        await asyncio.gather(
+            self._set_custom_permissions(structure["deep_branch"]),
+            self._set_custom_permissions(structure["sub_deep"]),
+            self._set_custom_permissions(structure["deep_files"][0]),
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async with detailed logging before dry run
+        await self._verify_list_acl_functionality(
+            entity=structure["deep_branch"],
+            expected_entity_count=3,  # deep_branch, sub_deep, and one deep_file
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I run delete_permissions with dry_run=True
+        await structure["deep_branch"].delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            dry_run=True,
+            show_acl_details=True,
+            show_files_in_containers=True,
+            synapse_client=self.syn_with_logger,
+        )
+
+        # THEN no permissions should be deleted
+        await asyncio.gather(
+            self._verify_permissions_not_deleted(structure["deep_branch"]),
+            self._verify_permissions_not_deleted(structure["sub_deep"]),
+            self._verify_permissions_not_deleted(structure["deep_files"][0]),
+        )
+
+        # AND comprehensive dry run analysis should be logged
+        self._verify_log_messages(
+            caplog,
+            list_acl_called=False,
+            delete_permissions_called=True,
+            dry_run=True,
+            tree_logging=False,
+        )
+
+        # Verify specific detailed logging messages
+        assert "DRY RUN: Permission Deletion Impact Analysis" in caplog.text
+        assert "End of Dry Run Analysis" in caplog.text
+
+    # Performance and stress tests
+    async def test_delete_permissions_large_flat_structure(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test deleting permissions on a large flat structure."""
+        # GIVEN a folder with many files
+        large_folder = await Folder(name=f"large_folder_{uuid.uuid4()}").store_async(
+            parent=project_model, synapse_client=self.syn
+        )
+        self.schedule_for_cleanup(large_folder.id)
+
+        # Create files in parallel
+        file_tasks = [
+            File(
+                path=utils.make_bogus_uuid_file(), name=f"large_file_{i}_{uuid.uuid4()}"
+            ).store_async(parent=large_folder, synapse_client=self.syn)
+            for i in range(10)  # Reduced from larger number for test performance
+        ]
+        files = await asyncio.gather(*file_tasks)
+
+        # Schedule cleanup for files
+        for file in files:
+            self.schedule_for_cleanup(file.id)
+
+        # Set permissions on all entities
+        entities_to_set = [large_folder] + files
+        await asyncio.gather(
+            *[self._set_custom_permissions(entity) for entity in entities_to_set]
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async performance with large structure
+        await self._verify_list_acl_functionality(
+            entity=large_folder,
+            expected_entity_count=11,  # large_folder + 10 files
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred for large structure
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions recursively
+        await large_folder.delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN all permissions should be deleted
+        await asyncio.gather(
+            *[self._verify_permissions_deleted(entity) for entity in entities_to_set]
+        )
+
+    async def test_delete_permissions_multiple_nested_branches(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test deleting permissions on multiple nested branches simultaneously."""
+        # GIVEN multiple complex nested branches
+        branch_tasks = [
+            Folder(name=f"branch_{branch_name}_{uuid.uuid4()}").store_async(
+                parent=project_model, synapse_client=self.syn
+            )
+            for branch_name in ["alpha", "beta"]
+        ]
+        branches = await asyncio.gather(*branch_tasks)
+
+        all_entities = list(branches)
+
+        # Schedule cleanup for branches
+        for branch in branches:
+            self.schedule_for_cleanup(branch.id)
+
+        # Create nested structure in each branch in parallel
+        nested_tasks = []
+        for branch_name, branch in zip(["alpha", "beta"], branches):
+            current_parent = branch
+            for level in range(2):
+                # Create sub-folder and file tasks for this level
+                sub_folder_task = Folder(
+                    name=f"{branch_name}_level_{level}_{uuid.uuid4()}"
+                ).store_async(parent=current_parent, synapse_client=self.syn)
+
+                nested_tasks.append(sub_folder_task)
+
+        # Execute all nested folder creation tasks in parallel
+        nested_folders = await asyncio.gather(*nested_tasks)
+
+        # Add nested folders to all_entities and schedule cleanup
+        all_entities.extend(nested_folders)
+        for folder in nested_folders:
+            self.schedule_for_cleanup(folder.id)
+
+        # Now create files for each nested folder in parallel
+        file_tasks = []
+        folder_index = 0
+        for branch_name in ["alpha", "beta"]:
+            for level in range(2):
+                parent_folder = nested_folders[folder_index]
+                file_task = File(
+                    path=utils.make_bogus_uuid_file(),
+                    name=f"{branch_name}_file_{level}_{uuid.uuid4()}",
+                ).store_async(parent=parent_folder, synapse_client=self.syn)
+                file_tasks.append(file_task)
+                folder_index += 1
+
+        # Execute all file creation tasks in parallel
+        files = await asyncio.gather(*file_tasks)
+
+        # Add files to all_entities and schedule cleanup
+        all_entities.extend(files)
+        for file in files:
+            self.schedule_for_cleanup(file.id)
+
+        # Set permissions on all entities
+        await asyncio.gather(
+            *[self._set_custom_permissions(entity) for entity in all_entities]
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async before deletion (complex multiple branches)
+        await self._verify_list_acl_functionality(
+            entity=project_model,
+            expected_entity_count=11,
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred for multiple branches
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions recursively from the project
+        await project_model.delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN all permissions should be deleted
+        await asyncio.gather(
+            *[self._verify_permissions_deleted(entity) for entity in all_entities]
+        )
+
+    async def test_delete_permissions_selective_branches(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test selectively deleting permissions from specific branches."""
+        # GIVEN multiple branches with permissions
+        # Create branches in parallel
+        branch_tasks = [
+            Folder(name=f"branch_a_{uuid.uuid4()}").store_async(
+                parent=project_model, synapse_client=self.syn
+            ),
+            Folder(name=f"branch_b_{uuid.uuid4()}").store_async(
+                parent=project_model, synapse_client=self.syn
+            ),
+        ]
+        branch_a, branch_b = await asyncio.gather(*branch_tasks)
+
+        # Schedule cleanup for branches
+        self.schedule_for_cleanup(branch_a.id)
+        self.schedule_for_cleanup(branch_b.id)
+
+        # Create files in each branch in parallel
+        file_tasks = [
+            File(
+                path=utils.make_bogus_uuid_file(), name=f"file_a_{uuid.uuid4()}"
+            ).store_async(parent=branch_a, synapse_client=self.syn),
+            File(
+                path=utils.make_bogus_uuid_file(), name=f"file_b_{uuid.uuid4()}"
+            ).store_async(parent=branch_b, synapse_client=self.syn),
+        ]
+        file_a, file_b = await asyncio.gather(*file_tasks)
+
+        # Schedule cleanup for files
+        self.schedule_for_cleanup(file_a.id)
+        self.schedule_for_cleanup(file_b.id)
+
+        # Set permissions on all entities
+        await asyncio.gather(
+            self._set_custom_permissions(branch_a),
+            self._set_custom_permissions(branch_b),
+            self._set_custom_permissions(file_a),
+            self._set_custom_permissions(file_b),
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async before selective deletion
+        await self._verify_list_acl_functionality(
+            entity=branch_a,
+            expected_entity_count=2,  # branch_a and file_a
+            recursive=True,
+            include_container_content=True,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred for selective branch
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions only from branch_a
+        await branch_a.delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN only branch_a and its contents should have permissions deleted
+        await asyncio.gather(
+            self._verify_permissions_deleted(branch_a),
+            self._verify_permissions_deleted(file_a),
+        )
+
+        # BUT branch_b should retain permissions
+        await asyncio.gather(
+            self._verify_permissions_not_deleted(branch_b),
+            self._verify_permissions_not_deleted(file_b),
+        )
+
+    async def test_delete_permissions_mixed_entity_types_in_structure(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test deleting permissions with mixed entity types in complex structure."""
+        # GIVEN a structure with both files and folders at multiple levels
+        structure = await self.create_complex_mixed_structure(project_model)
+
+        # Set permissions on a mix of entities
+        await asyncio.gather(
+            self._set_custom_permissions(structure["shallow_folder"]),
+            self._set_custom_permissions(structure["shallow_file"]),
+            self._set_custom_permissions(structure["deep_branch"]),
+            self._set_custom_permissions(structure["deep_files"][1]),
+            self._set_custom_permissions(structure["mixed_sub_folders"][0]),
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async with mixed entity types
+        await self._verify_list_acl_functionality(
+            entity=project_model,
+            expected_entity_count=5,  # All the entities we set permissions on
+            recursive=True,
+            include_container_content=True,
+            target_entity_types=["file", "folder"],
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred for mixed entity types
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions targeting both files and folders
+        await project_model.delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            target_entity_types=["file", "folder"],
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN all targeted entities should have permissions deleted
+        await asyncio.gather(
+            self._verify_permissions_deleted(structure["shallow_folder"]),
+            self._verify_permissions_deleted(structure["shallow_file"]),
+            self._verify_permissions_deleted(structure["deep_branch"]),
+            self._verify_permissions_deleted(structure["deep_files"][1]),
+            self._verify_permissions_deleted(structure["mixed_sub_folders"][0]),
+        )
+
+    async def test_delete_permissions_no_container_content_but_has_children(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test deleting permissions without include_container_content when children exist."""
+        # GIVEN a folder with children and custom permissions
+        parent_folder = await Folder(name=f"parent_folder_{uuid.uuid4()}").store_async(
+            parent=project_model, synapse_client=self.syn
+        )
+        self.schedule_for_cleanup(parent_folder.id)
+
+        child_file = await File(
+            path=utils.make_bogus_uuid_file(), name=f"child_file_{uuid.uuid4()}"
+        ).store_async(parent=parent_folder, synapse_client=self.syn)
+        self.schedule_for_cleanup(child_file.id)
+
+        # Set permissions on both entities
+        await asyncio.gather(
+            self._set_custom_permissions(parent_folder),
+            self._set_custom_permissions(child_file),
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async before testing container content exclusion
+        await self._verify_list_acl_functionality(
+            entity=parent_folder,
+            expected_entity_count=1,  # Only parent_folder, child excluded due to include_container_content=False
+            recursive=False,
+            include_container_content=False,
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred (should show limited structure)
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions without include_container_content
+        await parent_folder.delete_permissions_async(
+            include_self=True,
+            include_container_content=False,
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN only parent permissions should be deleted
+        await self._verify_permissions_deleted(parent_folder)
+
+        # AND child permissions should remain
+        await self._verify_permissions_not_deleted(child_file)
+
+    async def test_delete_permissions_case_insensitive_entity_types(
+        self, project_model: Project, caplog: pytest.LogCaptureFixture
+    ) -> None:
+        """Test that target_entity_types are case-insensitive."""
+        # GIVEN a simple structure with permissions
+        structure = await self.create_simple_tree_structure(project_model)
+        folder_a = structure["folder_a"]
+        file_1 = structure["file_1"]
+
+        # Set permissions on all entities
+        await asyncio.gather(
+            self._set_custom_permissions(folder_a),
+            self._set_custom_permissions(file_1),
+        )
+        await asyncio.sleep(2)
+
+        # WHEN - Verify list_acl_async with case-insensitive entity types
+        await self._verify_list_acl_functionality(
+            entity=folder_a,
+            expected_entity_count=2,  # folder_a and file_1 (case-insensitive filtering)
+            recursive=True,
+            include_container_content=True,
+            target_entity_types=[
+                "FOLDER",
+                "file",
+            ],  # Mixed case to test case-insensitivity
+            log_tree=True,
+        )
+
+        # Verify tree logging occurred with case-insensitive filtering
+        assert "ACL Tree Structure:" in caplog.text
+        caplog.clear()
+
+        # WHEN I delete permissions using mixed case entity types
+        await folder_a.delete_permissions_async(
+            recursive=True,
+            include_container_content=True,
+            target_entity_types=["FOLDER", "file"],  # Mixed case
+            dry_run=False,
+            synapse_client=self.syn,
+        )
+
+        # THEN all permissions should be deleted
+        await asyncio.gather(
+            self._verify_permissions_deleted(folder_a),
+            self._verify_permissions_deleted(file_1),
+        )
